Add project

+++ includes/read.h
@@ -8,8 +8,8 @@
 #ifndef READ_H_
     #define READ_H_
     #include "bsq.h"
-char *get_number(char * const str, size_t *number);
+char *get_number(char * const str, int *number);
 char *read_file_content(const char *file_path);
-size_t get_line_size(char const *data);
+int get_line_size(char const *data);
 int parse_file(const char *file_path, bsq_map *map);
 #endif /* !READ_H_ */
